Admin UI: refactored out session provider
The getSession() functionality wasn't used for anything, and as of 945ff089a60e5a1e1a8cdceb8df1b04f8d6263f4
isSignedIn wasn't either. This merges the mutations with the SignInPage/SignOutPage components and converts
their implementation to Apollo Hooks instead of using WithApollo.
