+++ packages/app-admin-ui/client/pages/List/dataHooks.js
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from 'react';
 import { __RouterContext } from 'react-router-dom';
 import debounce from 'lodash.debounce';
-import { useQuery } from 'react-apollo';
+import { useQuery } from '@apollo/react-hooks';
 
 import { deconstructErrorsToDataShape } from '../../util';
 import { pseudoLabelField } from './FieldSelect';
