+++ packages/app-admin-ui/package.json
@@ -8,6 +8,8 @@
     "node": ">=8.4.0"
   },
   "dependencies": {
+    "@apollo/react-components": "^3.1.3",
+    "@apollo/react-hooks": "^3.1.3",
     "@arch-ui/alert": "^0.0.8",
     "@arch-ui/badge": "^0.0.8",
     "@arch-ui/button": "^0.0.10",
@@ -70,7 +72,6 @@
     "prop-types": "^15.7.2",
     "raf-schd": "^4.0.0",
     "react": "^16.12.0",
-    "react-apollo": "^3.1.3",
     "react-document-title": "^2.0.3",
     "react-dom": "^16.12.0",
     "react-node-resolver": "^2.0.1",
